+++ CHANGELOG.md
@@ -7,6 +7,7 @@ HEAD
 * Fix `call of overloaded 'String(const char*, int)' is ambiguous`
 * Fix `JsonString` operator `==` and `!=` for non-zero-terminated string
 * Fix `-Wsign-conversion` on GCC 8 (issue #1715)
+* MessagePack: serialize round floats as integers (issue #1718)
 
 v6.19.2 (2022-02-14)
 -------